requirements.txt
torch
sentence-transformers
qdrant_client
openai
python-dotenv
tqdm